Addres another review of comments
